FIxed keys fix

+++ src/flexible_ddcm/solve.py
@@ -128,7 +128,7 @@ def get_continuation_value_for_transitions(
         continuation_values = np.zeros(transitions.shape)
     else:
         positions_continuation = state_space.variable_and_fixed_key_to_state[
-            np.ix_(fixed_keys, transitions.columns)
+            np.ix_(fixed_keys.astype(np.int32), transitions.columns)
         ]
         continuation_values = continuation_values.loc[positions_continuation].values
